Switch to new Jib CLI main and remove old main  (#2957)
