+++ jib-cli/src/main/java/com/google/cloud/tools/jib/cli/Build.java
@@ -17,144 +17,122 @@
 package com.google.cloud.tools.jib.cli;
 
 import com.google.cloud.tools.jib.api.Containerizer;
-import com.google.cloud.tools.jib.api.DockerDaemonImage;
-import com.google.cloud.tools.jib.api.ImageReference;
-import com.google.cloud.tools.jib.api.Jib;
-import com.google.cloud.tools.jib.api.JibContainer;
 import com.google.cloud.tools.jib.api.JibContainerBuilder;
-import com.google.cloud.tools.jib.api.LogEvent;
-import com.google.cloud.tools.jib.api.buildplan.AbsoluteUnixPath;
-import com.google.cloud.tools.jib.api.buildplan.FileEntriesLayer;
-import com.google.cloud.tools.jib.api.buildplan.Port;
-import com.google.cloud.tools.jib.cli.JibCli.ImageReferenceParser;
+import com.google.cloud.tools.jib.api.LogEvent.Level;
+import com.google.cloud.tools.jib.cli.buildfile.BuildFiles;
+import com.google.cloud.tools.jib.cli.logging.CliLogger;
+import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLogger;
+import com.google.cloud.tools.jib.plugins.common.logging.SingleThreadedExecutor;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Joiner;
-import java.util.List;
-import java.util.Map.Entry;
+import com.google.common.base.Verify;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.Duration;
+import java.util.Collections;
+import java.util.Map;
 import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.function.Consumer;
-import javax.annotation.Nullable;
-import picocli.CommandLine.Command;
-import picocli.CommandLine.Parameters;
+import picocli.CommandLine;
+import picocli.CommandLine.Model.CommandSpec;
 
-/** A Jib CLI subcommand for building a container image from a set of files on disk. */
-@Command(name = "build", description = "Create a container image from static files")
-public class Build extends Building implements Callable<Integer> {
+@CommandLine.Command(
+    name = "build",
+    showAtFileInUsageHelp = true,
+    description = "Build a container")
+public class Build implements Callable<Integer> {
 
-  @Parameters(
-      index = "0",
-      paramLabel = "base-image",
-      description = "The base image (ex: busybox, nginx, gcr.io/distroless/java)",
-      converter = ImageReferenceParser.class)
-  @VisibleForTesting
+  @CommandLine.Spec
   @SuppressWarnings("NullAway.Init") // initialized by picocli
-  ImageReference baseImage;
+  private CommandSpec spec;
 
-  @Parameters(
-      index = "1",
-      paramLabel = "destination-image",
-      description =
-          "The destination image (ex: localhost:5000/image:1.0, gcr.io/project/image:latest)",
-      converter = ImageReferenceParser.class)
-  @VisibleForTesting
+  @CommandLine.Mixin
   @SuppressWarnings("NullAway.Init") // initialized by picocli
-  ImageReference destinationImage;
-
-  @Parameters(
-      index = "2..*",
-      paramLabel = "layer-spec",
-      description =
-          "Create a layer from local file-system. A layer-spec "
-              + "is a set of mappings of the form:\n"
-              + "    local/path[,/container/path[,directive1,...]]\n"
-              + "Container path defaults to '/' if omitted.\n"
-              + "Directives include:\n"
-              + "\n"
-              + "  name=<xxx>    set the layer name\n"
-              + "  p=perms       set file and directory permissions:\n"
-              + "    actual      use actual values in file-system\n"
-              + "    <fff>:<ddd> octal file and directory permissions\n"
-              + "  ts=timestamp  set last-modified timestamps:\n"
-              + "    actual      use actual values in file-system\n"
-              + "    <number>    seconds since Unix epoch\n"
-              + "    <xxx>       date-time in ISO8601 format\n"
-              + "\n"
-              + "Default permissions are 644 for files and 755 for directories. "
-              + "Default timestamps are 1970-01-01 00:00:01 UTC. "
-              + "Multiple mappings may be specified, separated by a semi-colon ';'.",
-      converter = LayerDefinitionParser.class)
   @VisibleForTesting
-  @Nullable
-  List<FileEntriesLayer> layers;
+  CommonCliOptions commonCliOptions;
+
+  @CommandLine.Option(
+      names = {"-c", "--context"},
+      defaultValue = ".",
+      paramLabel = "<project-root>",
+      description = "The context root directory of the build (ex: path/to/my/build/things)")
+  @SuppressWarnings("NullAway.Init") // initialized by picocli
+  private Path contextRoot;
+
+  @CommandLine.Option(
+      names = {"-b", "--build-file"},
+      paramLabel = "<build-file>",
+      description = "The path to the build file (ex: path/to/other-jib.yaml)")
+  @SuppressWarnings("NullAway.Init") // initialized by picocli
+  private Path buildFile;
+
+  @CommandLine.Option(
+      names = {"-p", "--parameter"},
+      paramLabel = "<name>=<value>",
+      description =
+          "templating parameter to inject into build file, replace $${<name>} with <value> (repeatable)")
+  private Map<String, String> templateParameters = Collections.emptyMap();
+
+  public Path getContextRoot() {
+    return Verify.verifyNotNull(contextRoot);
+  }
+
+  /**
+   * Returns a user configured Path to a buildfile and if none is configured returns jib.yaml in
+   * {@link #getContextRoot()}.
+   *
+   * @return a path to a bulidfile
+   */
+  public Path getBuildFile() {
+    if (buildFile == null) {
+      return getContextRoot().resolve("jib.yaml");
+    }
+    return buildFile;
+  }
+
+  public Map<String, String> getTemplateParameters() {
+    return templateParameters;
+  }
 
   @Override
-  public Integer call() throws Exception {
-    Consumer<LogEvent> logger = System.out::println;
-    JibContainerBuilder builder = Jib.from(parent.toCredentialedImage(baseImage, logger));
-    verbose("FROM " + baseImage);
-    if (creationTime != null) {
-      builder.setCreationTime(creationTime);
-    }
-    if (entrypoint != null) {
-      verbose("ENTRYPOINT [" + Joiner.on(",").join(entrypoint) + "]");
-      builder.setEntrypoint(entrypoint);
-    }
-    if (arguments != null) {
-      verbose("CMD [" + Joiner.on(",").join(arguments) + "]");
-      builder.setProgramArguments(arguments);
-    }
-    if (environment != null) {
-      for (Entry<String, String> pair : environment.entrySet()) {
-        verbose("ENV " + pair.getKey() + "=" + pair.getValue());
-        builder.addEnvironmentVariable(pair.getKey(), pair.getValue());
-      }
-    }
-    if (labels != null) {
-      for (Entry<String, String> pair : labels.entrySet()) {
-        verbose("LABEL " + pair.getKey() + "=" + pair.getValue());
-        builder.addLabel(pair.getKey(), pair.getValue());
-      }
-    }
-    if (ports != null) {
-      for (Port port : ports) {
-        verbose("EXPOSE " + port);
-        builder.addExposedPort(port);
-      }
-    }
-    if (volumes != null) {
-      for (AbsoluteUnixPath volume : volumes) {
-        verbose("VOLUME " + volume);
-        builder.addVolume(volume);
-      }
-    }
-    if (user != null) {
-      verbose("USER " + user);
-      builder.setUser(user);
-    }
-    if (layers != null) {
-      for (FileEntriesLayer layer : layers) {
-        builder.addFileEntriesLayer(layer);
-      }
-    }
-    Containerizer containerizer =
-        pushMode.toDocker
-            ? Containerizer.to(DockerDaemonImage.named(destinationImage))
-            : Containerizer.to(parent.toCredentialedImage(destinationImage, logger));
-    containerizer.setAllowInsecureRegistries(parent.insecure);
-    containerizer.setToolName("jib");
-    containerizer.addEventHandler(LogEvent.class, logger);
-
-    ExecutorService executor = Executors.newCachedThreadPool();
+  public Integer call() {
+    commonCliOptions.validate();
+    SingleThreadedExecutor executor = new SingleThreadedExecutor();
     try {
-      containerizer.setExecutorService(executor);
+      ConsoleLogger logger =
+          CliLogger.newLogger(
+              commonCliOptions.getVerbosity(),
+              commonCliOptions.getConsoleOutput(),
+              spec.commandLine().getOut(),
+              spec.commandLine().getErr(),
+              executor);
 
-      JibContainer result = builder.containerize(containerizer);
-      System.out.printf("Containerized to %s (%s)\n", destinationImage, result.getDigest());
-      return 0;
+      if (!Files.isReadable(buildFile)) {
+        logger.log(
+            Level.ERROR,
+            "The Build File YAML either does not exist or cannot be opened for reading: "
+                + buildFile);
+        return 1;
+      }
+      if (!Files.isRegularFile(buildFile)) {
+        logger.log(Level.ERROR, "Build File YAML path is a not a file: " + buildFile);
+        return 1;
+      }
+
+      CacheDirectories cacheDirectories = CacheDirectories.from(commonCliOptions, contextRoot);
+      Containerizer containerizer = Containerizers.from(commonCliOptions, logger, cacheDirectories);
+
+      JibContainerBuilder containerBuilder =
+          BuildFiles.toJibContainerBuilder(contextRoot, buildFile, this, commonCliOptions, logger);
+
+      containerBuilder.containerize(containerizer);
+    } catch (Exception ex) {
+      if (commonCliOptions.isStacktrace()) {
+        ex.printStackTrace();
+      }
+      System.err.println(ex.getClass().getName() + ": " + ex.getMessage());
+      return 1;
     } finally {
-      executor.shutdown();
+      executor.shutDownAndAwaitTermination(Duration.ofSeconds(3));
     }
+    return 0;
   }
 }
