+++ jib-cli/src/main/java/com/google/cloud/tools/jib/cli/Containerizers.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.cli.cli2;
+package com.google.cloud.tools.jib.cli;
 
 import static com.google.cloud.tools.jib.api.Jib.DOCKER_DAEMON_IMAGE_PREFIX;
 import static com.google.cloud.tools.jib.api.Jib.REGISTRY_IMAGE_PREFIX;
