+++ jib-cli/src/main/java/com/google/cloud/tools/jib/cli/logging/CliLogger.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.cli.cli2.logging;
+package com.google.cloud.tools.jib.cli.logging;
 
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLogger;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLoggerBuilder;
