+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/CredentialsTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.cli.cli2;
+package com.google.cloud.tools.jib.cli;
 
 import static com.google.common.truth.Truth.assertThat;
