+++ jib-cli/src/test/java/com/google/cloud/tools/jib/cli/logging/CliLoggerTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.cli.cli2.logging;
+package com.google.cloud.tools.jib.cli.logging;
 
 import static com.google.common.truth.Truth.assertThat;
